requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.4.0
pydantic>=2.0.0
datasets>=2.14.0
git-python>=1.0.0
docker>=6.1.0
openai>=1.0.0
subprocess32>=3.5.4 